microservicio4/requirements.txt
fastapi
uvicorn
httpx[http2]
python-dotenv
requests
